backend/foodgram/requirements.txt
asgiref==3.6.0
django==3.2.14
django-filter==22.1
django-extra-fields==3.0.2
django-colorfield==0.8.0
djoser==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytest-drf==1.1.3
Pillow==9.4.0
isort==5.11.4
#flake8==6.0.0
gunicorn==20.1.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3